gpt_pdf_md/requirements.txt
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
google-api-core==2.13.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
idna==3.4
load-dotenv==0.1.0
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==10.1.0
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypandoc==1.12
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
urllib3==2.0.7
